Small clean-up
